Move stuff out of _private header
